CMake fix

+++ Caelis Chaos/src/CaelisEngine/Utilities/Util.h
@@ -13,7 +13,7 @@ inline SDL_FPoint SDLFPoint(float x, float y) {
     return p;
 }
 
-inline SDL_Color SDLColor(int r, int g, int b, int a) {
+inline SDL_Color SDLColor(Uint8 r, Uint8 g, Uint8 b, Uint8 a) {
     SDL_Color c = {r, g, b, a};
     return c;
 }
